Added filter functionaliy , Add category , Delete posts, Comment

+++ src/post/post.controller.ts
@@ -13,17 +13,20 @@ import {
   UseGuards,
   UseInterceptors,
   Res,
+  NotFoundException,
 } from '@nestjs/common';
 import { PostService } from './post.service';
-import {User_} from '../user/userv2.decorator'
+import { User_ } from '../user/userv2.decorator';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { AuthGuard } from '@nestjs/passport';
-import { Request } from 'express';
-import { User } from '../User/entities/user.entity';
 import { ACGuard, UseRoles } from 'nest-access-control';
 import { FileInterceptor } from '@nestjs/platform-express';
 import multer, { diskStorage } from 'multer';
+import { UserRoles } from 'src/models/user-roles.models';
+import { Roles } from 'src/user/user.roles.decorator';
+import { Post as PostEntity } from './entities/post.entity';
+import { RolesGuard } from 'src/user/user.roles.guard';
 
 @Controller('posts')
 @UseInterceptors(ClassSerializerInterceptor)
@@ -32,16 +35,10 @@ export class PostController {
 
   @Post()
   @UseGuards(AuthGuard('jwt'), ACGuard)
-  @UseRoles({
-    resource: 'posts',
-    action: 'create',
-    possession: 'any',
-  })
   create(@Body() createPostDto: CreatePostDto, @User_() user) {
     return this.postService.create(createPostDto, user);
   }
 
-  // Upload Picture to Server
   @Post('upload-photo')
   @UseInterceptors(
     FileInterceptor('picture', {
@@ -76,7 +73,6 @@ export class PostController {
     }
   }
 
-  //  GET Photos
   @Get('pictures/:fileId')
   async serveAvatar(@Param('fileId') fileId, @Res() res): Promise<any> {
     res.sendFile(fileId, { root: './uploads' });
@@ -84,7 +80,6 @@ export class PostController {
 
   @Get()
   findAll(@Query() query: any) {
-    // We are only targeting query parameters, slug and sort
     return this.postService.findAll(query);
   }
 
@@ -110,13 +105,14 @@ export class PostController {
   }
 
   @Delete(':id')
-  @UseGuards(AuthGuard('jwt'), ACGuard)
-  @UseRoles({
-    resource: 'posts',
-    action: 'delete',
-    possession: 'any',
-  })
-  remove(@Param('id') id: string) {
+  @UseGuards(AuthGuard('jwt'), RolesGuard)
+  @Roles(UserRoles.Reader)
+  remove(
+    @Param('id') id: string,
+  ): Promise<{ success: boolean; post: PostEntity }> {
+    if (!this.postService.findOne(+id)) {
+      throw new NotFoundException('Could not find the post to delete');
+    }
     return this.postService.remove(+id);
   }
 }
